requirements.txt
Django==5.0
djangorestframework==3.14.0
environ==1.0
django-environ==0.11.2
drf-yasg==1.21.7
pip==23.3.1
requests~=2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.1